Update 3-optimize-network-performance.md
Lots of updates here. I took out the ORM stuff as I think it's going to be over the head of many, and we don't really explain what it is in detail to really connect the dots (nor do I want to). I've also added a section on ExpressRoute. I'd like to add one on accelerated networking but may do that post-Ignite. Unless you have any other updates, let's get this merged. We'll still need images and questions.