M31410: Missing word
Hello, @markjulmar,
Localization team has reported source content issue that causes issue for localization. 
Description: It seems there is a missing term after "the name to the" in this sentence. Could you please advise what is it referring to?
Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.Many thanks in advance.