Dev/copy edits + make it easy
Did a text and tone pass. Also, modified the painful parts of this module to be less painful. For example, using a Custom Extension Script to make it easy to configure the VM (versus manual SSH in and running many commands.) Another example is using `echo | tee` instead of walking the user through every keystroke using `vi`.