M32069: Possible error: "cannot" should be "can"
Hello, @jboeshart, @markjulmar,
Translator has reported possible source content issue. 
Description: Please confirm if the "cannot" on the following sentence should be "can", otherwise the sentence does not make sense. We translated as if it was "can".
Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.