Revision
* Lots of voice and tone edits.
* Consolidate 11 units into 7 by combining knowledge needed with the task.
* Rename titles for simplicity.
* Add additional context and links to other resources.
* Add task validation.
* Likely lots of other stuff. This is somewhat lengthy change.

What got me started on this module was my desire to make it easier for the learner to go through. The manual SSH and SCP stuff makes it a bit laborious. Ideally, I would want to show how to build the web app locally, and then how to deploy it to Azure (either manually or through a script.) Unfortunately, you can't install software on Cloud Shell, so we're limited as to what we can do here.

Resolves #1072
Resolves #2904