Publish new module - Build Azure VM templates (#3152)
* Add scaffolding

* Drafting

* Set title

* Add badge for "Build Azure virtual machine templates"

* Specify correct includes

* Tweaks

* Drafting

* Add images

* Drafting

* Drafting

* Drafting

* Tweak

* Split into pivots

* Move pivot declarations to main page

* Fix media refs

* Drafting

* Add task validation

* Add knowledge check questions

* Set title and level

* Rename image

It was named after the wrong unit.

* Add "Build Azure Resource Manager templates" to LP

* Set alt text

* Add border

* Rename media folder

* Set sandbox <rgn> tag

* Set metadata

* Set description
* Reduce quiz indentation
* Set duration
* Set summary

* Run Acrolinx

* Tweak summary

* Add standard sandbox includes

* Remove location section

The location is not relevant - it's tied to the parent resource group, which we provide.

* Tweak duration

* Set summary

* Cleanup URLs

Remove the locale

* Open links in new tab

* s/QuickStart/Quickstart

* Copy edits

* Suggestions (#3187)

* Copy edits

Nick provided some really good feedback on the Linux pivot. This change applies the same changes to the Windows pivot.

* Clarify term and soften callout

* More precisely define Azure Resource Manager
* Make the tip about VS Code feel less like an "ad"

* Add sandbox region

I want to show the RG, but there was a regression in displaying them in output boxes.
Tracking https://ceapex.visualstudio.com/CEINTL/_workitems/edit/55867 before we publish this.

* Tweak write/deploy guidance

* Simplify write guiance.
* Move deploy guidance closer to the task.

* Rename unit

"Add a resource to the template" is much clearer than "Build a template resource".

* Remove output tag

This fixes a bug where the <rgn> tag wasn't getting translated to the resource group name.

* build-azure-vm-templates suggestions (#3193)

* suggested updates to unit 2

* unit 3 updates

* updates to unit 4 and unit 2 (from @nickwalkmsft other PR)

* update to unit 5

* Number the procedures

For clarity

* Remove extra whitespace

* Add bash language to code block

As a test, see if it causes the <rgn> tag to render.

* Set badge

* Add scaffolding

* Drafting

* Set title

* Add badge for "Build Azure virtual machine templates"

* Specify correct includes

* Tweaks

* Drafting

* Add images

* Drafting

* Drafting

* Drafting

* Tweak

* Split into pivots

* Move pivot declarations to main page

* Fix media refs

* Drafting

* Add task validation

* Add knowledge check questions

* Set title and level

* Rename image

It was named after the wrong unit.

* Add "Build Azure Resource Manager templates" to LP

* Set alt text

* Add border

* Rename media folder

* Set sandbox <rgn> tag

* Set metadata

* Set description
* Reduce quiz indentation
* Set duration
* Set summary

* Run Acrolinx

* Tweak summary

* Add standard sandbox includes

* Remove location section

The location is not relevant - it's tied to the parent resource group, which we provide.

* Tweak duration

* Set summary

* Cleanup URLs

Remove the locale

* Open links in new tab

* s/QuickStart/Quickstart

* Copy edits

* Suggestions (#3187)

* Copy edits

Nick provided some really good feedback on the Linux pivot. This change applies the same changes to the Windows pivot.

* Clarify term and soften callout

* More precisely define Azure Resource Manager
* Make the tip about VS Code feel less like an "ad"

* Add sandbox region

I want to show the RG, but there was a regression in displaying them in output boxes.
Tracking https://ceapex.visualstudio.com/CEINTL/_workitems/edit/55867 before we publish this.

* Tweak write/deploy guidance

* Simplify write guiance.
* Move deploy guidance closer to the task.

* Rename unit

"Add a resource to the template" is much clearer than "Build a template resource".

* Remove output tag

This fixes a bug where the <rgn> tag wasn't getting translated to the resource group name.

* build-azure-vm-templates suggestions (#3193)

* suggested updates to unit 2

* unit 3 updates

* updates to unit 4 and unit 2 (from @nickwalkmsft other PR)

* update to unit 5

* Number the procedures

For clarity

* Remove extra whitespace

* Add bash language to code block

As a test, see if it causes the <rgn> tag to render.

* Set badge

* Rename file