Run Docker containers with ACI - Remove unused Bash variable (#3242)
* Remove unused Bash variable

This addresses [ceapex 58012](https://ceapex.visualstudio.com/CEINTL/_workitems/edit/58012).

The feedback was that `STORAGE_ACCOUNT` variable is created but not used, since you already have that information. The fix here is to remove creating the variable. The step that creates that variable is not useful because it relies on the other variable, which already contains the value.

Extras:

* Rename Bash variable `ACI_PERS_STORAGE_ACCOUNT_NAME` to `STORAGE_ACCOUNT_NAME` for simplicity.
* Fix a typo

* Fix variable reference

* Remove $