M33465: Incomplete sentence (#3324)
Hello @markjulmar, @jboeshart,
Translator has reported possible source content issue.
Description: Source sentence seems to be problematic due to missing words (...when resources 'are created'...). We localized the string as if the words 'are created' place as shown.
Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.