Recommended updated example IP address (#3590)
Just a recommendation to change the IP address in the `for example` text to match the image in the next image. I've noticed novices use pattern matching, so even though the IP address is just an example, it could be confusing to have the example output of running the `az vm show` command be different than the image showing the url where the webpage has navigated to.