Update 5-add-web-server.md (#3533)
There was a problem with the syntax for a few of the code samples. The samples worked in Bash, but PowerShell is a little less-forgiving with regard to how it parses single-quote and double-quote characters. In order to get the command to work in PowerShell, I reversed all of the single-quote and double-quote characters, and now the examples work fine in both PowerShell and Bash.