Update 7-exercise-using-the-azure-portal-to-enable-logging-and-download-logs.md
Logging is not enabled by default, so I had to add that section. Otherwise this exercise did not work at all, which took a great deal of time to troubleshoot and discover what was going wrong. (Note that there is no mention of using the **Tables/$MetricsHourPrimaryTransactionsBlob** log, which is pretty handy.) Also, the Azure Storage Explorer was updated, so a lot of that section had to be rewritten.