Updating Create and run stored procedures from your app (#4053)
* Updating Create and run stored procedures from your app

Based on the description, I believe that the original author intended to include a different stored procedure and never added it to the unit. However, if you remove the description, the rest of the unit is correct.

* Update 6-stored-procedure.md

Added some additional verbiage, which was really just to reset the verification process since it appeared to have frozen.