115511: replacing italic asterisk by <em> tag
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.
Description: Reporting source issues as a result of nested asterisks (i.e. *** instead of **_) that are breaking code in localized sites. Although, for this case we couldn't apply the underscore fix because it got displayed as plain text, so we changed it by a <em> tag