Update 4-exercise-setup-communication.md
@nickwalkmsft for awareness

The previous code that installed the azure iot cli extension needed to be updated to current. 

For an explanation and example in a current Azure article, see:
https://docs.microsoft.com/en-us/azure/iot-hub/quickstart-send-telemetry-cli#prepare-two-cli-sessions

Note that the Azure CLI team notified us of the above bug in this article, they scan the repos looking for outdated instances of the azure iot cli extension.