M50470: Moving 2015 to avoid localization issues
Translator has reported possible source content issue. 
Translator reported missing content from localization files since the "2015." part is in a new line and the parser tool, detect it as an isolated number. As isolated numbers are not localizables, it is not available for translation and for some languages it needs to be placed in a different part of the string.