Update 10-string-methods.md
Removed duplication 0  This is the first time we've seen methods, but methods are not unique to
strings. As we will see shortly, other types of objects can have them. have them. 

Corrected to 
This is the first time we've seen methods, but methods are not unique to
strings. As we will see shortly, other types of objects can have them.