Removed DevOps learning objective and prereq
Among the learning objectives there's "Determine strategy for platform automation and DevOps". However, I could not find any content related to DevOps in this module. For consistency, I would suggest to either add a new unit about DevOps design area to this module or to remove the DevOps learning objective and prerequisite.