Update 13-how-azure-rtos-environment.md
Line 29: Remove "en-us/"
Line 43: Remove "en-au/" from this URL.
Line 7, 23, 41, etc. (throughout the module): This line before the heading is a nonstandard design element for Learn. Could you remove it? If not, we will need to get admin approval before we can merge from this URL to make it locale agnostic.