Update 3-contribute.md
Was checking where the exercise fits and might have suggested a couple of text updates ;) You've covered everything I wanted to cover in the exercise, but this: writing a commit message when there isn't clear guidance on conventions. Maybe for the exercise I can do an open ended question for folks to ponder over what to write in such a case... 

For this module, maybe include reasons why a maintainer might decide against merging your PR? Like, when it's not in the scope of what the maintainer is comfortable to ... well... maintain, and when it adds complexity while solving a very niche use case only?