Update 6-exercise-add-a-durable-timer-to-manage-a-long-running-task.md
IMPORTANT NOTE: I split the library installation commands into two separate steps, because when I was going through this module, I used Learn's "Copy" feature to copy the combined commands and pasted them into the Console window according to the module steps. However, the second command to install the _moment_ library didn't take, which made the sample code fail, and I spent a couple hours debugging that before I discovered the problem. With that in mind, it seemed like a better idea to split the commands so that they are run independently, and therefore less prone to error.