Performed renames as requested
Corrected the following

Remove "and" from the folder/module name /intro-to-data-protection-and-privacy/. If this is impractical, explain in a comment.
Thanks. #hold-off

Following are non-blocking issues that should be resolved now or in another PR within three business days:

Remove "and" from the file names 4-common-data-protection-regulations-and-acts.yml/.md and 5-how-microsoft-support-protection-and-privacy.yml/.md (four file names).