Update 3-how-azure-vmware-solution-works.md
update NSX-T version to the correct version supported (https://docs.microsoft.com/en-us/azure/azure-vmware/concepts-private-clouds-clusters#vmware-software-versions)

As per PM (Amit Aneja): "We have a mix in production today. We are deploying 2.5.2 now by default. The reason we have a mix we are because we upgrade some to 3.1.1 and hit a regression and that's why we stopped 3.1.1 deployment."

Version 3.1.1 is being bypassed. The next version (3.1.2) will be deployed in July sometime.

update HCX to latest version