Update 5-initiate-live-response-session.md
The table looks good (OK) in the Preview. I have tried creating "new paragraph marks" for ending the lines before and after the header file of the table (line 60), and before and after  the table. This is the 3rd commit and I don't see the error on my connection.Thanks.