Update 3-exercise-configure-auditing.md
Adding a new line into the first Powershell command so that when users choose Copy and paste into the Azure Cloud Shell it will automatically run the command. We have found users have struggled when pasting commands to hit the ENTER key so get confused on why the command just doesn't execute. This change is a test to see if that works by editing the first command in this exercise to add an empty newline