Refresh simple-website learn module
This learn module is getting a bit old and I thought it could use a refresher.
It also lacked more links to DevTools docs, as knowing how to use the Elements and Console panel is really important when you start web development.
I thought it was a pitty to keep this only for the end, so I sprinkled a few links.
I also found a few confusing parts that I tried to clarify.
Finally, I updated all screenshots which had a few problems:
- some were showing Chrome DevTools, while it's a good idea to use this opportunity to show our own Edge DevTools
- some didn't show enough context, so I replaced them with the full window, and added red outlines, to make it easier for people to understand what to look at and where to find things
- I also replaced the custom ::: image tagging with actual markdown tagging. The custom tagging wasn't adding any value apart from a border around the images. But since I took screen captures of full windows, with the nice windows 11 rounded borders, we don't need those anymore. Using markdown tagging makes it simpler to read, and maintain, and also works when previewed in the github UI.