Update 4-interacting-data.md
Missed this in my previous PR, but ran into it on the live stream. EnsureCreated returns false because we have been using Migrations to create the database, and hence it already exists. However, we still want to seed it, so we should run `Initialize` method anyway. It already checks that the database does not have data before doing anything, so it is safe to run every time.