Updating Text-to-Speech to lowercase
Unfortunately, I have seen inconsistent situations where "Text-to-Speech" is used as part of the product name, and other situations where "text-to-speech" is used to describe a generic technology. In this case, I have no problems with erring on the side of using "text-to-speech" is used to describe a generic technology.