Incorporated Review Changes
Line 57: Remove HTML/replace HTML with Markdown (<p>, <ul>, <li>). <br> can be used in tables (only) to create lists. To create a list in a table, use a hyphen (-) instead of each bullet point. Separate list items by using the line-break HTML tag <br>. Note that the limitation of this workaround is that it doesn't create actual bullet points in the rendered article.