Update index.yml
Capitalizing the 'b' in Build in the title to be consistent with other guided project titles. In addition, the capitalization of the first word after "guided project" is necessary to improve the title's appearance in search results. Often, search engines chop off the "Guided project" prefix and begin with the first word after the hyphen.