Update 5-exercise-configure-resource-lock.md
Removing link to a training module "Create an Azure account" because referrals coming from this training module are causing low completion rates in the "Create an Azure account" module. Instructional Design guidance recommends not linking to modules from other training modules. I've replaced the link to the module with a link to the sign up page.