Fixed blocking issues
3-multi-tenant-model.md

The list in lines 31-36 should be bulleted, not numbered.
5-microservices-on-azure.md

Same as above for lists in lines 7-13, 56-61, and 67-71.
create_approle.png
ent_app.png
media/expose_api.png

Replace underscore with hyphen in these file names.
6-authentification-and-authorization.md

Remove comma in line 79 and use the Learn standard formatting for notes.
ent_app.png

Use a red triangle for emphasis as is Learn standard.
expose_api.png

Replace the blacked out info with appropriate fictional info or the background color.
3-multi-tenant-model.yml (Words and Phrases)
3-multi-tenant-model.md (Words and Phrases and Correctness)
5-microservices-on-azure.md (Words and Phrases and Correctness)
6-authentification-and-authorization.md (Words and Phrases and Correctness)
8-summary.md (Words and Phrases)

The articles above need Acrolinx-requested fixes in the cited categories. (See the individual Acrolinx scorecards for details.)
6-authentification-and-authorization.md

In lines 33 and 43 and 83, begin with the type of graphic (for example, "Screenshot of..." or "Diagram that shows...") as for the alt text. (And delete "The following".)
6-authentification-and-authorization.yml

Lowercase "Authorization" and "Solutions" in meta title (same in subheads and text throughout, but that is not blocking).
index.yml

Lowercase "Foundations" in meta title.